moved plugins under plugins dir
